Correct data type of series save_no field

+++ Sma5h/Mods/Sma5h.Mods.Music/Models/SeriesEntry.cs
@@ -9,7 +9,7 @@ namespace Sma5h.Mods.Music.Models
         public string NameId { get; set; }
         public sbyte DispOrder { get; set; }
         public sbyte DispOrderSound { get; set; }
-        public short SaveNo { get; set; }
+        public sbyte SaveNo { get; set; }
         public bool Unk1 { get; set; }
         public bool IsDlc { get; set; }
         public bool IsPatch { get; set; }
